drivers: misc: pn547: Implements dynamic allocation for frame size
Implements dynamic allocation for frame size by using kcalloc and
avoid unnecessary #ifdefs.

Why kcalloc?
The advantage of this one is that will prevent overflows
and it is also a bit nicer to read.

Also, this patch checks the ioctl state (power on with[out] firmware)
for all devices which prevents possible wakelocks.

Signed-off-by: Humberto Borba <humberos@gmail.com>
Change-Id: Ie8beecf39c1a76cc9a05951f8e94661842c72429